WebResult can be absent, in that case the WebResult name should be "return" and targetNamespace should be that of service.
Removed the exception thrown if there is no wrapper parameter found in the wrapper bean, that can be the case with jaxb generated bean. The assumption at this point should be to default to the result QName to the default values and not throw exceptions.

This fixes most of the regression caused in test failures.