Issue number: Obtained from: Submitted by: Reviewed by:
Changing the publshing pattern to wsdl=1, xsd=1 instead of wsdl=a.wsdl.
wsdl=a.wsdl is possible because all the wsdls and schemas are in /WEB-INF/wsdl
in J2EE. The generation of ?wsdl=a.wsdl cannot be done for J2SE endpoint as
systemids don't have any pattern. The change will make deployment consistent
across J2EE and J2SE.