Issue number: Obtained from: Submitted by: Reviewed by:
Bug fix: 6286071
When dispatch method is not found, it raises correct exception with env:Client.
But that is not converted to the wire message correctly. In this case
javamethod was null while converting MessageInfo to InternalMessage
and this was used to find the bindingId. Now messageInfo is used to find the
bindingId and that is used to create correct fault message.