introduced ByteArrayBuffer class, which is much like ByteArrayOutputStream but without its performance problem.
Ideally, buffering of data should not be done anywhere in the stack, but if you have to, use this instead of
ByteArrayOutputStream. I replaced most of the existing use of ByteArrayOutputStream with this new class.

We should eventually look into why so many buffering is done inside the RI, but that's to be done later.