Removed domSourceToNode method.
Inside the JAX-WS RI, we are mostly converting to DOM only to add it to something else (mostly SAAJ.)
If we build a new DOM from scratch and then move a tree into another document, a whole tree gets copied
and as a result you pay a severe performance penalty.

This convenience method was therefore not desirable for use inside the runtime.