Issue number: Obtained from: Submitted by: Reviewed by:
Bug fix : 6336275

Sending HTTP stats code 400 for malformed xml request. It's difficult to find
whether the XML is malformed or not as JAXB throws exception UnmarshallException
(cause seems to be XMLStreamException). The XMLStreamException may be lost or get
burried under different layers. Put a fix that looks in exception chain and tries
to send Bad Request status code.