added Header.getAttribute() as requested by the security team.
Since this can be used to implement methods like getRole(), isRelay(), and so on,
I added AbstractHeaderImpl that does that. For most of the outbound messages, such
methods will never be invoked, so it makes sense for messages like JAXBMessage and
SourceMessage to use this to reduce the code size.