reworked the Packet.isOneWay property.
The ReliableMessaging team identified that there can be a disconnect between
what the calling application expects (whether it's one way or not) and what
the transport should be doing. To reconcile these difference in such a way
that makes it easy for pipes to check for one-way messages, I changed
the isOneWay property.


  - Packet.isOneWay property is removed. To be precise, it's left there,
    with @deprecated, so that the removal does not cause immediate build
    failure for Tango. But this flag is no longer updated, used,
    referenced, and will be really removed shortly.
  
  - A part of the functionality is moved as Message.isOneWay(WSDLPort).
    This method lets pipes check whether a message is a one-way request
    according to the WSDL. This is done by checking the tag name of the
    payload.
  
  - There is a new property called Message.expectReply, which instructs
    one-way transports (like SMTP) whether or not wait for a reply message.


RM will want to set Message.expectReply=true, to request the
transport to check for a reply. It can also check if the current request
message is a one-way request message, by using Message.isOneWay(WSDLPort)