this code is now used by ProtocolSourceMessage to potentially read from a DOM tree, so making an assumption that strings are interned is not safe anymore.
I'm torn on this. The majority of the case is to read from stream, so it makes sense
to rely on interning to make that faster.

OTOH, doing String.intern in DOM XMLStreamReader seems to be unnecessarily expensive
(so does a filter.)

For now this is blocking a WCF databinding interop test, so I'll fix it the easy way.
This could use some discussion.