Fix for bug 6447399: service.getPort(sei) throws NPE, if the service is craeted using Service API instead of generated service.
Changing lot of things.
Annotations on the generated service class (@WebServicesClient, @WebEndpoint) are not used at all to guess or figure out anything while building the model. 

getPort(sei), Since PortQName is not passed, First port in the wsdl with matching portType is used instead of parsing the @WebEndpoint to get the PortQName.

Need to do some refactoring ( remove  unused methods).
Will fail, if two Ports use the same PortType.