Bug fixes for CR 6448733 and CR 6448728. This should bring down a lot of failures in SQE worksapce.
Correct faultCode is used when there is a PE.
When there is a PE or handleMessage(0 returns false on client, it should close only logical handlers, it should not try to close cousinPipe as it is not called.
Using a flag requestProcessingSuccessful to decide whether to close cousinPipe or not while calling cousinPipe.close(). Here assuming that cousinPipe will be called later if request processing is sucessful in this Pipe.