[NEWS:for-vijay for-ryan] ResourceInjector abstraction added. This newly introduced abstraction allows the JAX-WS RI to drive the resource injection process, instead of other way around (which is what GF has been doing.)
This is a crucial piece for stateful web service support, as it allows the JAX-WS RI to manage multiple instances of the service endpoint, while still conforming to the EE semantics.

Once this is integrated into Glassfish, we can have Glassfish's NewJAXWSServlet fixed. Since ResourceInjector is responsible for injecting WebServiceContext, and it's given JAX-WS RI's WebServiceContext implementation, this enables Glassfish to perform a WebServiceContext filtering without needing to use Invoker.