Issue number: Obtained from: Submitted by: Reviewed by: Fix for https://wsit.dev.java.net/issues/show_bug.cgi?id=232
If a war file contains multiple WSDL/XSDs, a mex request for endpoint1 could
also include WSDL/XSD for endpoint2. This is fixed by finding the transitive
closure of the documents for each endpoint. WSEndpoint.getServiceDefinition()
will give only that endpoint's documents. This means some customization
files may require to be changed since previously the mapping is based on the
entire war. This is not a big change since it is mostly changing xsd=3 to xsd=2.