oporting the fix from 2.1.3  thats assists wsit fix https://wsit.dev.java.net/issues/show_bug.cgi?id=706.
Don't add wsa:Action if there is no SEIModel or WSDLModel. This expects the user to set the wsa:Action header on the message and will not add a default action header.  This would be helpful for Provider endpoints to set their own action header on the response and still use Addressing Tube functionality.